refactor: 타입 정의 types.ts 파일에 분리
