+++ client/src/Components/Input.tsx
@@ -1,20 +1,9 @@
 import React, { forwardRef } from 'react';
 import styled from 'styled-components';
 import color from '../util/color';
+import { InputProps } from '../types';
 
 const { brown, bordergrey } = color;
-
-interface InputProps {
-  type: string;
-  readOnly?: boolean;
-  placeholder: string;
-  paddingRight?: string;
-  marginBottom?: string;
-  width?: string;
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  openAddressModal?: (e: React.MouseEvent<HTMLInputElement>) => void;
-  onKeyUp?: (e: React.KeyboardEvent<HTMLInputElement>) => void;
-}
 
 const Input = (
   {
