+++ client/src/Components/SearchBar.tsx
@@ -2,12 +2,9 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import Swal from 'sweetalert2';
 import color from '../util/color';
+import { SearchBarProps } from '../types';
 
 const { yellow, darkbrown, brown } = color;
-
-interface SearchBarProps {
-  search: (searchType: string, searchContent: string) => void;
-}
 
 function SearchBar({ search }: SearchBarProps) {
   const [type, setType] = useState<string>('author');
