+++ client/src/Components/SortModal.tsx
@@ -1,13 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
 import color from '../util/color';
+import { SortModalProps } from '../types';
 
 const { ivory, darkivory, yellow, brown } = color;
-
-interface SortModalProps {
-  setSorting: React.Dispatch<React.SetStateAction<string>>;
-  setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
-}
 
 function SortModal({ setSorting, setIsOpen }: SortModalProps) {
   const clickHandler = (text: string) => {
